# resource: Optimize integrity string generation