@bep 2018-10-16
Sure, but note that a 250ns save like this is only worth while in a very hot loop, and in most cases I would keep the fmt version for clarity. But in this case, the new code actually looks better.